Moved base64 and hmacsha1 functions to a sharable location
