+++ Proxima/mpconfigport.h
@@ -74,7 +74,8 @@
 #define MICROPY_PY_UJSON                            (1)
 #define MICROPY_PY_URE                              (1)
 #define MICROPY_PY_UHEAPQ                           (1)
-#define MICROPY_PY_UHASHLIB                         (0)
+#define MICROPY_PY_UHASHLIB                         (1)
+#define MICROPY_PY_UHASHLIB_HMACSHA1                (1)
 
 #define MICROPY_ENABLE_EMERGENCY_EXCEPTION_BUF      (1)
 #define MICROPY_EMERGENCY_EXCEPTION_BUF_SIZE        (0)
