+++ lib/crypto/hmacsha1.c
